+++ src/math/Matrix4x4.h
@@ -4,22 +4,22 @@
 
 class Matrix4x4
 {
-	double values[16];
+	Float values[16];
 public:
 
 	inline	Matrix4x4();
 	inline	Matrix4x4(const Matrix4x4&);
-	inline 	Matrix4x4(	double, double, double, double,
-						double, double, double, double, 
-						double, double, double, double, 
-						double, double, double, double);
+	inline 	Matrix4x4(	Float, Float, Float, Float,
+						Float, Float, Float, Float, 
+						Float, Float, Float, Float, 
+						Float, Float, Float, Float);
 
-					double& 	operator[](unsigned idx)		{return values[idx];}
-	const			double& 	operator[](unsigned idx) const	{return values[idx];}
-					double& 	operator()(unsigned i, unsigned j)			{return values[i*4+j];}
-	const			double& 	operator()(unsigned i, unsigned j) const 	{return values[i*4+j];}
+					Float& 	operator[](unsigned idx)		{return values[idx];}
+	const			Float& 	operator[](unsigned idx) const	{return values[idx];}
+					Float& 	operator()(unsigned i, unsigned j)			{return values[i*4+j];}
+	const			Float& 	operator()(unsigned i, unsigned j) const 	{return values[i*4+j];}
 
-	inline			Matrix4x4	operator* (double);
+	inline			Matrix4x4	operator* (Float);
 	inline			Matrix4x4	operator* (const Matrix4x4&);
 	inline const	Matrix4x4	operator* (const Matrix4x4&) const;
 	inline			Matrix4x4	operator+ (const Matrix4x4&);
@@ -28,12 +28,12 @@ public:
 	inline const	Matrix4x4 t() const;
 					Matrix4x4 inv(); //To implement
 
-			double* begin()			{return &values[0];}
-	const	double* begin()	const	{return &values[0];}
-			double* end()			{return &values[16];}
-	const 	double* end()	const	{return &values[16];}
+			Float* begin()			{return &values[0];}
+	const	Float* begin()	const	{return &values[0];}
+			Float* end()			{return &values[16];}
+	const 	Float* end()	const	{return &values[16];}
 };
-inline Matrix4x4 operator*(double a, Matrix4x4 m) {return m*a;}
+inline Matrix4x4 operator*(Float a, Matrix4x4 m) {return m*a;}
 
 #include "Matrix4x4.inl"
